Service completo com JPA

+++ src/main/java/edu/ifam/dra/chat/controller/ContatoController.java
@@ -1,6 +1,5 @@
 package edu.ifam.dra.chat.controller;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,7 +34,7 @@ public class ContatoController {
 	}
 	
 	@GetMapping("/{id}")
-	ResponseEntity<Contato> getContato(@PathVariable int id) {
+	ResponseEntity<Contato> getContato(@PathVariable Long id) {
 		try {
 			return ResponseEntity.ok(contatoService.getContato(id));
 		} catch(Exception e) {
@@ -51,7 +50,7 @@ public class ContatoController {
 	}
 	
 	@PutMapping("/{id}")
-	ResponseEntity<Contato> setContato(@RequestBody Contato contato, @PathVariable int id) {
+	ResponseEntity<Contato> setContato(@RequestBody Contato contato, @PathVariable Long id) {
 		try {
 			return ResponseEntity.accepted().body(contatoService.setContato(id, contato));
 		} catch(Exception e) {
@@ -61,7 +60,7 @@ public class ContatoController {
 	}
 	
 	@DeleteMapping("/{id}")
-	ResponseEntity<Contato> deleteContato(@PathVariable int id) {
+	ResponseEntity<Contato> deleteContato(@PathVariable Long id) {
 		try {
 			contatoService.deleteContato(id);
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new Contato());
